// src/types/data.ts
export interface Project {
  title: string
  description: string
  link?: string
  github: string
  image: string
}

export interface Experience {
  jobTitle: string
  company: string
  companyUrl: string
  startDate: string
  endDate: string
  location: string
  description: string
}

export interface Contact {
  name: string
  href: string
}